% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathead.R
\name{mathead}
\alias{mathead}
\title{head of a matrix}
\usage{
mathead(x, ...)
}
\arguments{
\item{x}{the matrix to print}

\item{...}{ignored}
}
\value{
the original matrix (invisibl)
}
\description{
similar to `head()` but does not attempt to print big matrices
}
\examples{
mat <- matrix(runif(50*60), nrow = 50)
rownames(mat) <- c("qwertyuiopasdfghjkl", "123456789123456789") |> rep(25)
mathead(mat)
}
