% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matimage.R
\name{matimage}
\alias{matimage}
\title{Quick plotting of matrix}
\usage{
matimage(mat, ...)
}
\arguments{
\item{mat}{the matrix to plot}

\item{...}{parameters passed to `image`}
}
\value{
the original matrix invisibly
}
\description{
Plot a matrix using `image` with raster and replotting axes. Adapted from `imageM()` in the examples of `?image`.
}
